Modularize scenario step validation and server capabilities
Move compile-time step validation into cfg-gated helpers in scenarios/step_validation.rs, replacing the cfg_if! selection. Extract build_server_capabilities into server/server_capabilities.rs to keep server.rs under the module-length lint. Expose INDEXING_COUNTER to sibling tests via pub(super). Deduplicate the RFC section in docs/contents.md and drop the unused tracing dev-dependency from rstest-bdd. Validated with make check-fmt, lint, typecheck, and test (118 passed).
