+++ crates/rstest-bdd-server/src/server.rs
@@ -20,22 +20,17 @@ use crate::{
         WorkspaceRoot,
         index_feature_source_owned,
     },
-    lsp::{
-        ClientCapabilities,
-        ServerCapabilities,
-        TextDocumentSyncCapability,
-        TextDocumentSyncKind,
-        TextDocumentSyncOptions,
-        WorkspaceFolder,
-    },
+    lsp::{ClientCapabilities, WorkspaceFolder},
 };
 
 mod deferred_saves;
 mod scenario_scopes;
+mod server_capabilities;
 mod workspace_task;
 
 use deferred_saves::{DeferredDocumentSaves, DeferredSaveDropReason};
 use scenario_scopes::ScenarioScopeRegistry;
+pub use server_capabilities::build_server_capabilities;
 use workspace_task::WorkspaceTask;
 
 /// Central state shared across all LSP handlers.
@@ -376,33 +371,6 @@ impl ServerState {
     pub fn remove_rust_step_index(&mut self, path: &Path) {
         self.rust_step_indices.remove(path);
         self.step_registry.invalidate_file(path);
-    }
-}
-
-/// Build the server capabilities to advertise to the client.
-///
-/// Phase 7 advertises text document sync to receive save notifications for
-/// `.feature` file indexing, definition navigation for Rust-to-feature step
-/// navigation, and implementation navigation for feature-to-Rust step
-/// navigation.
-#[must_use]
-pub fn build_server_capabilities() -> ServerCapabilities {
-    ServerCapabilities {
-        text_document_sync: Some(TextDocumentSyncCapability::Options(
-            TextDocumentSyncOptions {
-                open_close: Some(true),
-                change: Some(TextDocumentSyncKind::INCREMENTAL),
-                save: Some(crate::lsp::TextDocumentSyncSaveOptions::SaveOptions(
-                    crate::lsp::SaveOptions {
-                        include_text: Some(true),
-                    },
-                )),
-                ..Default::default()
-            },
-        )),
-        definition_provider: Some(crate::lsp::OneOf::Left(true)),
-        implementation_provider: Some(crate::lsp::ImplementationProviderCapability::Simple(true)),
-        ..ServerCapabilities::default()
     }
 }
