+++ crates/rstest-bdd-server/src/server/scenario_scopes.rs
@@ -17,8 +17,6 @@ use crate::indexing::{CompiledStepDefinition, IndexedScenarioBinding, ScenarioBi
 pub(super) struct ScenarioScopeRegistry {
     /// Bindings replaced atomically when one Rust file is re-indexed.
     bindings_by_file: HashMap<PathBuf, Vec<IndexedScenarioBinding>>,
-    /// Memoised scope resolutions, cleared whenever any bindings change.
-    resolved_scopes: HashMap<PathBuf, Result<Option<Vec<String>>, ScenarioScopeConflict>>,
 }
 
 /// Conflicting closed scopes selected by bindings for one feature file.
@@ -57,7 +55,6 @@ impl ScenarioScopeRegistry {
         } else {
             self.bindings_by_file.insert(path.to_path_buf(), bindings);
         }
-        self.resolved_scopes.clear();
     }
 
     /// Return the single closed scope selected for one feature file.
@@ -65,24 +62,17 @@ impl ScenarioScopeRegistry {
         &self,
         feature_path: &Path,
     ) -> Result<Option<Vec<String>>, ScenarioScopeConflict> {
-        if let Some(cached) = self.resolved_scopes.get(feature_path) {
-            return cached.clone();
-        }
         let scopes: BTreeSet<_> = self
             .matching_bindings(feature_path)
             .map(|binding| normalized_scope(&binding.libraries))
             .collect();
-        let resolved = if scopes.len() > 1 {
-            Err(ScenarioScopeConflict {
+        if scopes.len() > 1 {
+            return Err(ScenarioScopeConflict {
                 feature_path: feature_path.to_path_buf(),
                 scopes: scopes.into_iter().collect(),
-            })
-        } else {
-            Ok(scopes.into_iter().next())
-        };
-        self.resolved_scopes
-            .insert(feature_path.to_path_buf(), resolved.clone());
-        resolved
+            });
+        }
+        Ok(scopes.into_iter().next())
     }
 
     /// Iterate over bindings whose target contains one feature file.
