+++ docs/contents.md
@@ -31,14 +31,6 @@
   limitations that maintainers should keep visible.
 - [Changelog](CHANGELOG.md) records repository-level release history.
 
-
-## Requests for comments
-
-- [RFC 0001: Explicit step-library scopes][rfc-0001]
-  defines closed-world, lexical selection of a scenario's step vocabulary.
-- [RFC 0002: Named step-argument binding][rfc-0002]
-  defines name-based `StepArgs` capture binding and shared conversion policy.
-
 ## Requests for comments
 
 - [RFC 0001: Explicit step-library scopes][rfc-0001]
